chore: remove participant list from space view
