+++ lib/pages/chat_list/space_view.dart
@@ -21,7 +21,6 @@ import 'package:fluffychat/pangea/extensions/pangea_room_extension.dart';
 import 'package:fluffychat/pangea/public_spaces/public_room_bottom_sheet.dart';
 import 'package:fluffychat/pangea/spaces/constants/space_constants.dart';
 import 'package:fluffychat/pangea/spaces/widgets/knocking_users_indicator.dart';
-import 'package:fluffychat/pangea/spaces/widgets/leaderboard_participant_list.dart';
 import 'package:fluffychat/utils/localized_exception_extension.dart';
 import 'package:fluffychat/utils/stream_extension.dart';
 import 'package:fluffychat/widgets/adaptive_dialogs/show_ok_cancel_alert_dialog.dart';
@@ -855,14 +854,6 @@ class _SpaceViewState extends State<SpaceView> {
                     //   },
                     // ),
                     KnockingUsersIndicator(room: room),
-                    SliverList.builder(
-                      itemCount: 1,
-                      itemBuilder: (context, i) {
-                        return LeaderboardParticipantList(
-                          space: room,
-                        );
-                      },
-                    ),
                     // Pangea#
                     SliverList.builder(
                       itemCount: joinedRooms.length,
